% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-broom.R
\name{add_index}
\alias{add_index}
\title{Adds a sequential index column to a data frame}
\usage{
add_index(ret, index_rename)
}
\arguments{
\item{ret}{An object of class tibble}

\item{index_rename}{A variable indicating the index name to be used in the
tibble returned}
}
\description{
Adds a sequential index column to a data frame
}
