% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-forecast.R
\name{arima_string}
\alias{arima_string}
\title{Print the ARIMA model parameters}
\usage{
arima_string(object, padding = FALSE)
}
\arguments{
\item{object}{An object of class Arima}

\item{padding}{Add padding to the name returned}
}
\description{
Refer to forecast:::arima.string.
\href{https://github.com/robjhyndman/forecast/blob/master/R/arima.R}{forecast arima.R}
}
