% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-forecast.R
\name{bats_string}
\alias{bats_string}
\title{Print the BATS model parameters}
\usage{
bats_string(object)
}
\arguments{
\item{object}{An object of class bats}
}
\description{
Refer to forecast:::makeText.
\href{https://github.com/robjhyndman/forecast/blob/master/R/bats.R}{forecast bats.R}
}
