% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-bike_sharing.R
\docType{data}
\name{bike_sharing_daily}
\alias{bike_sharing_daily}
\title{Daily Bike Sharing Data}
\format{
A tibble: 731 x 16
\itemize{
\item instant: record index
\item dteday : date
\item season : season (1:winter, 2:spring, 3:summer, 4:fall)
\item yr : year (0: 2011, 1:2012)
\item mnth : month ( 1 to 12)
\item hr : hour (0 to 23)
\item holiday : weather day is holiday or not (extracted from \link{Web Link})
\item weekday : day of the week
\item workingday : if day is neither weekend nor holiday is 1, otherwise is 0.
\item weathersit :
\itemize{
\item 1: Clear, Few clouds, Partly cloudy, Partly cloudy
\item 2: Mist + Cloudy, Mist + Broken clouds, Mist + Few clouds, Mist
\item 3: Light Snow, Light Rain + Thunderstorm + Scattered clouds, Light Rain + Scattered clouds
\item 4: Heavy Rain + Ice Pallets + Thunderstorm + Mist, Snow + Fog
}
\item temp : Normalized temperature in Celsius. The values are derived via (t-t_min)/(t_max-t_min), t_min=-8, t_max=+39 (only in hourly scale)
\item atemp: Normalized feeling temperature in Celsius. The values are derived via (t-t_min)/(t_max-t_min), t_min=-16, t_max=+50 (only in hourly scale)
\item hum: Normalized humidity. The values are divided to 100 (max)
\item windspeed: Normalized wind speed. The values are divided to 67 (max)
\item casual: count of casual users
\item registered: count of registered users
\item cnt: count of total rental bikes including both casual and registered
}
}
\usage{
bike_sharing_daily
}
\description{
This dataset contains the daily count of
rental bike transactions between years 2011 and 2012 in Capital bikeshare
system with the corresponding weather and seasonal information.
}
\examples{
bike_sharing_daily

}
\references{
Fanaee-T, Hadi, and Gama, Joao,
'Event labeling combining ensemble detectors and background knowledge',
Progress in Artificial Intelligence (2013): pp. 1-15, Springer Berlin Heidelberg,
\href{https://link.springer.com/article/10.1007/s13748-013-0040-3}{Web Link}.
}
\keyword{datasets}
