% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-condense_by_time.R
\name{condense_by_time}
\alias{condense_by_time}
\title{Condense the Period (e.g. Daily to Monthly)}
\usage{
condense_by_time(.data, .date_var, .by = "day", .type = c("first", "last"))
}
\arguments{
\item{.data}{A \code{tbl} object or \code{data.frame}}

\item{.date_var}{A column containing date or date-time values to summarize.
If missing, attempts to auto-detect date column.}

\item{.by}{A time unit to aggregate by.
Time units are collapsed using \code{lubridate::floor_date()} or \code{lubridate::ceiling_date()}.

The value can be:
\itemize{
\item \code{second}
\item \code{minute}
\item \code{hour}
\item \code{day}
\item \code{week}
\item \code{month}
\item \code{bimonth}
\item \code{quarter}
\item \code{season}
\item \code{halfyear}
\item \code{year}
}

Arbitrary unique English abbreviations as in the \code{lubridate::period()} constructor are allowed:
\itemize{
\item \code{"1 year"}
\item \code{"2 months"}
\item \code{"30 seconds"}
}}

\item{.type}{One of "first" or "last".}
}
\value{
A \code{tibble} or \code{data.frame}
}
\description{
Convert a \code{data.frame} object from daily to monthly,
from minute data to hourly, and more. This allows the user to easily
aggregate data to a less granular level by taking the value from either
the beginning or end of the period.
}
\examples{
# Libraries
library(timetk)
library(dplyr)

# First value in each month
m4_daily \%>\%
    group_by(id) \%>\%
    condense_by_time(.by = "1 month")

# Last value in each month
m4_daily \%>\%
    group_by(id) \%>\%
    condense_by_time(.by = "1 month", .side = "end")



}
\seealso{
Time-Based dplyr functions:
\itemize{
\item \code{\link[=summarise_by_time]{summarise_by_time()}} - Easily summarise using a date column.
\item \code{\link[=filter_by_time]{filter_by_time()}} - Quickly filter using date ranges.
\item \code{\link[=between_time]{between_time()}} - Range detection for date or date-time sequences.
\item \code{\link[=pad_by_time]{pad_by_time()}} - Insert time series rows with regularly spaced timestamps
\item \code{\link[=condense_by_time]{condense_by_time()}} - Convert to a different periodicity
\item \code{\link[=slidify]{slidify()}} - Turn any function into a sliding (rolling) function
}
}
