% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-diff.R
\name{diff_vec}
\alias{diff_vec}
\alias{diff_inv_vec}
\title{Differencing Transformation}
\usage{
diff_vec(.x, .lag = 1, .difference = 1, .log = FALSE)

diff_inv_vec(
  .x,
  .lag = 1,
  .difference = 1,
  .log = FALSE,
  .initial_values = NULL
)
}
\arguments{
\item{.x}{A numeric vector to be differenced or inverted.}

\item{.lag}{Which lag (how far back) to be included in the differencing calculation.}

\item{.difference}{The number of differences to perform.
\itemize{
\item 1 Difference is equivalent to period change.
\item 2 Differences is equivalent to acceleration.
}}

\item{.log}{If log differences should be calculated.
\emph{Note that difference inversion of a log-difference is approximate.}}

\item{.initial_values}{A numeric vector of the initial values, which are used for difference inversion.
This vector is the original values that are the length of the \code{NA} missing differences.}
}
\value{
A numeric vector
}
\description{
\code{diff_vec()} applies a Differencing Transformation.
\code{diff_inv_vec()} inverts the differencing transformation.
}
\details{
\strong{Benefits:}

This function is \code{NA} padded by default so it works well with \code{dplyr::mutate()} operations.
}
\examples{
library(dplyr)
library(timetk)

# --- MUTATE ----
m4_daily \%>\%
    group_by(id) \%>\%
    mutate(difference = diff_vec(value, .lag = 1)) \%>\%
    mutate(
        difference_inv = diff_inv_vec(
            difference,
            .lag        = 1,
            # Add initial value to calculate the inverse difference
            .initial_values = value[1]
        )
    )

}
\seealso{
\itemize{
\item Rolling Window Calculations: \code{\link[=roll_apply_vec]{roll_apply_vec()}}, \code{\link[slider:slide_vec]{slider::slide_vec()}}, \code{\link[zoo:rollapply]{zoo::rollapply()}}
\item Loess Modeling: \code{\link[stats:loess]{stats::loess()}}
}
}
