% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dates.R
\name{get_date_variables}
\alias{get_date_variables}
\title{Get date or date-time variables}
\usage{
get_date_variables(df)
}
\arguments{
\item{df}{An object of class data.frame}
}
\description{
Refer to padr:::get_date_variables.
\href{https://github.com/EdwinTh/padr/blob/master/R/helpers.R}{padr helpers.R}
}
