% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-impute_ts.R
\name{impute_ts_vec}
\alias{impute_ts_vec}
\title{Missing Value Imputation for Time Series}
\usage{
impute_ts_vec(x, period = 1, lambda = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{period}{A seasonal period to use during the transformation. If \code{period = 1},
linear interpolation is performed. If \code{period > 1}, a robust STL decomposition is
first performed and a linear interpolation is applied to the seasonally adjusted data.}

\item{lambda}{A box cox transformation parameter. If set to \code{"auto"}, performs
automated lambda selection.}
}
\description{
This is mainly a wrapper for the Seasonally Adjusted Missing Value Interpolation function,
\code{na.interp()}, from the \code{forecast} R package. This function includes arguments for applying
seasonality to numeric vector (non-\code{ts}) via the \code{period} argument.
}
\details{
\strong{Linear Interpolation}

With \code{period = 1}, the algorithm uses linear interpolation.

\strong{Seasonal Interpolation}

For seasonal series with \code{period > 1}, a robust Seasonal Trend Loess (STL) decomposition is first computed.
Then a linear interpolation is applied to the seasonally adjusted data, and
the seasonal component is added back.
}
\examples{
library(dplyr)
library(timetk)


# --- VECTOR ----

values <- c(1,2,3, 4*2, 5,6,7, NA, 9,10,11, 12*2)
values

# Linear interpolation
impute_ts_vec(values, period = 1, lambda = NULL)

# Seasonal Interpolation: set period = 4
impute_ts_vec(values, period = 4, lambda = NULL)

# Seasonal Interpolation with Box Cox Transformation (internal)
impute_ts_vec(values, period = 4, lambda = "auto")


}
\references{
\itemize{
\item \href{https://github.com/robjhyndman/forecast}{Forecast R Package}
\item \href{https://otexts.com/fpp2/missing-outliers.html}{Forecasting Principles & Practices: Dealing with missing values and outliers}
}
}
\seealso{
\itemize{
\item Box Cox Transformation: \code{\link[=box_cox_vec]{box_cox_vec()}}
\item Lag Transformation: \code{\link[=lag_vec]{lag_vec()}}
\item Differencing Transformation: \code{\link[=diff_vec]{diff_vec()}}
\item Rolling Window Transformation: \code{\link[=roll_apply_vec]{roll_apply_vec()}}
\item Loess Smoothing Transformation: \code{\link[=smooth_vec]{smooth_vec()}}
\item Fourier Series: \code{\link[=fourier_vec]{fourier_vec()}}
\item Missing Value Imputation for Time Series: \code{\link[=impute_ts_vec]{impute_ts_vec()}}
}
}
