% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dates.R
\name{is_date_class}
\alias{is_date_class}
\title{Check if an object is a date class}
\usage{
is_date_class(x)
}
\arguments{
\item{x}{A vector to check}
}
\value{
Logical (TRUE/FALSE)
}
\description{
Check if an object is a date class
}
\examples{
library(dplyr)

tk_make_timeseries("2011") \%>\% is_date_class()

letters \%>\% is_date_class()


}
