% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-m4_monthly.R
\docType{data}
\name{m4_monthly}
\alias{m4_monthly}
\title{Sample of 4 Monthly Time Series Datasets from the M4 Competition}
\format{
A tibble: 9,743 x 3
\itemize{
\item \code{id} Factor. Unique series identifier (4 total)
\item \code{date} Date. Timestamp information. Monthly format.
\item \code{value} Numeric. Value at the corresponding timestamp.
}
}
\source{
\itemize{
\item \href{https://mofc.unic.ac.cy/m4/}{M4 Competition Website}
}
}
\usage{
m4_monthly
}
\description{
The fourth M Competition.
M4, started on 1 January 2018 and ended in 31 May 2018.
The competition included 100,000 time series datasets.
This dataset includes \strong{a sample of 4 monthly time series from the competition.}
}
\details{
This is a sample of 4 Monthly data sets from the M4 competition.
}
\examples{
m4_monthly

}
\keyword{datasets}
