% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-m4_quarterly.R
\docType{data}
\name{m4_quarterly}
\alias{m4_quarterly}
\title{Sample of 4 Quarterly Time Series Datasets from the M4 Competition}
\format{
A tibble: 9,743 x 3
\itemize{
\item \code{id} Factor. Unique series identifier (4 total)
\item \code{date} Date. Timestamp information. Quarterly format.
\item \code{value} Numeric. Value at the corresponding timestamp.
}
}
\source{
\itemize{
\item \href{https://www.unic.ac.cy/iff/research/forecasting/m-competitions/m4/}{M4 Competition Website}
}
}
\usage{
m4_quarterly
}
\description{
The fourth M Competition.
M4, started on 1 January 2018 and ended in 31 May 2018.
The competition included 100,000 time series datasets.
This dataset includes \strong{a sample of 4 quarterly time series from the competition.}
}
\details{
This is a sample of 4 Quarterly data sets from the M4 competition.
}
\examples{
m4_quarterly

}
\keyword{datasets}
