% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-m4_weekly.R
\docType{data}
\name{m4_weekly}
\alias{m4_weekly}
\title{Sample of 4 Weekly Time Series Datasets from the M4 Competition}
\format{
A tibble: 9,743 x 3
\itemize{
\item \code{id} Factor. Unique series identifier (4 total)
\item \code{date} Date. Timestamp information. Weekly format.
\item \code{value} Numeric. Value at the corresponding timestamp.
}
}
\source{
\itemize{
\item \href{https://mofc.unic.ac.cy/m4/}{M4 Competition Website}
}
}
\usage{
m4_weekly
}
\description{
The fourth M Competition.
M4, started on 1 January 2018 and ended in 31 May 2018.
The competition included 100,000 time series datasets.
This dataset includes \strong{a sample of 4 weekly time series from the competition.}
}
\details{
This is a sample of 4 Weekly data sets from the M4 competition.
}
\examples{
m4_weekly

}
\keyword{datasets}
