% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lag_diagnostics.R
\name{plot_lag_diagnostics}
\alias{plot_lag_diagnostics}
\title{Plot Lag Diagnostics: ACF, PACF, and CCF}
\usage{
plot_lag_diagnostics(
  .data,
  .value,
  ...,
  .lags = 0:20,
  .facet_ncol = 1,
  .facet_scales = "fixed",
  .line_color = "#2c3e50",
  .line_size = 0.5,
  .point_color = "#2c3e50",
  .point_size = 1,
  .hline_color = "#2c3e50",
  .title = "Lag Diagnostics",
  .x_lab = "Lag",
  .y_lab = "Correlation",
  .interactive = TRUE
)
}
\arguments{
\item{.data}{A data frame or tibble with numeric features (values) in descending
chronological order}

\item{.value}{A numeric column with a value to have ACF and PACF calculations
performed.}

\item{...}{Additional features to perform Lag Cross Correlations (CCFs)
versus the \code{.value}. Useful for evaluating external lagged regressors.}

\item{.lags}{A seqence of one or more lags to evaluate.}

\item{.facet_ncol}{Facets: Number of facet columns. Has no effect if using \code{grouped_df}.}

\item{.facet_scales}{Facets: Options include "fixed", "free", "free_y", "free_x"}

\item{.line_color}{Line color. Use keyword: "scale_color" to change the color by the facet.}

\item{.line_size}{Line size}

\item{.point_color}{Point color. Use keyword: "scale_color" to change the color by the facet.}

\item{.point_size}{Point size}

\item{.hline_color}{Color for the y-intercept = 0 line.}

\item{.title}{Title for the plot}

\item{.x_lab}{X-axis label for the plot}

\item{.y_lab}{Y-axis label for the plot}

\item{.interactive}{Returns either a static (\code{ggplot2}) visualization or an interactive (\code{plotly}) visualization}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
\code{plot_lag_diagnostics()} returns the ACF and PACF of a
target and optionally any CCF's of lagged predictors.
The lag visualization is available in static (\code{ggplot2}) and
interactive (\code{plotly}) formats.
}
\details{
\strong{Simplified ACF, PACF, & CCF}

We are often interested in all 3 of these functions. Why not get all 3 at once?
Now you can!
\itemize{
\item \strong{ACF} - Autocorrelation between a target variable and lagged versions of itself
\item \strong{PACF} - Partial Autocorrelation removes the dependence of lags on
other lags highlighting key seasonalities.
\item \strong{CCF} - Shows how lagged predictors can be used for prediction of a target
variable.
}

\strong{Works with Grouped Data Frames}

The \code{tk_lag_diagnostics()} works with \code{grouped_df}'s, meaning you can
group your time series by one or more categorical columns with \code{group_by()}
and then apply \code{tk_lag_diagnostics()} to return group-wise lag diagnostics.
}
\examples{
library(tidyverse)
library(tidyquant)
library(timetk)


# Plotly - Interactive Visualization (Good for Exploration)
FANG \%>\%
    group_by(symbol) \%>\%
    plot_lag_diagnostics(
        .value  = adjusted,  # ACF & PACF
        volume, close,       # CCF
        .lags   = 0:180
    )

# ggplot2 - static visualization (Good for PDF Reports)
FANG \%>\%
    group_by(symbol) \%>\%
    plot_lag_diagnostics(
        .value  = adjusted,  # ACF & PACF
        volume, close,       # CCF
        .lags   = 0:180,
        .interactive = FALSE
    )


}
\seealso{
\itemize{
\item \code{\link[=tk_lag_diagnostics]{tk_lag_diagnostics()}}: Function that returns ACF, PACF, & CCF data
\item \code{\link[=plot_time_series]{plot_time_series()}}: A useful interactive plotting function
}
}
