% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-time_series.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot a Time Series}
\usage{
plot_time_series(
  .data,
  .date_var,
  .value,
  .facets = NULL,
  .facet_ncol = 1,
  .facet_scales = "free_y",
  .line_color = "#2c3e50",
  .line_size = 0.5,
  .y_intercept = NULL,
  .y_intercept_color = "#2c3e50",
  .smooth = TRUE,
  .smooth_period = NULL,
  .smooth_degree = 2,
  .smooth_color = "#3366FF",
  .smooth_size = 1,
  .title = "Time Series Plot",
  .x_lab = "",
  .y_lab = "",
  .interactive = TRUE
)
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame}}

\item{.date_var}{A column containing either date or date-time values}

\item{.value}{A column containing numeric values}

\item{.facets}{One or more grouping columns that broken out into \code{ggplot2} facets.
These can be selected using \code{tidyselect()} helpers (e.g \code{contains()}).}

\item{.facet_ncol}{Facets: Number of facet columns.}

\item{.facet_scales}{Facets: Options include "fixed", "free", "free_y", "free_x"}

\item{.line_color}{Line color. Use keyword: "scale_color" to change the color by the facet.}

\item{.line_size}{Line size}

\item{.y_intercept}{Value for a y-intercept on the plot}

\item{.y_intercept_color}{Color for the y-intercept}

\item{.smooth}{Logical - Whether or not to include a trendline smoother.
Uses See \code{\link[=smooth_vec]{smooth_vec()}} to apply a LOESS smoother.}

\item{.smooth_period}{Number of observations to include in the Loess Smoother.
See \code{\link[=smooth_vec]{smooth_vec()}}.}

\item{.smooth_degree}{Flexibility of Loess Polynomial.
Either 0, 1, 2 (0 = lest flexible, 2 = more flexible).}

\item{.smooth_color}{Smoother line color}

\item{.smooth_size}{Smoother line size}

\item{.title}{Title for the plot}

\item{.x_lab}{X-axis label for the plot}

\item{.y_lab}{Y-axis label for the plot}

\item{.interactive}{Returns either a static (\code{ggplot2}) visualization or an interactive (\code{plotly}) visualization}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
\code{plot_time_series()} returns a time series visualization in either static (\code{ggplot2}) or
interactive (\code{plotly}) formats.
}
\examples{
library(tidyverse)
library(tidyquant)
library(timetk)


# Plotly - Interactive Visualization (Good for Exploration)
FANG \%>\%
    plot_time_series(date, adjusted, .facets = contains("symbol"))

# ggplot2 - static visualization (Good for PDF Reports)
FANG \%>\%
    plot_time_series(
         date, adjusted,
        .facets        = symbol,
        .facet_ncol    = 2,
        .line_color    = "scale_color",
        .smooth_period = 180,
        .interactive   = FALSE) +
   theme_tq_dark() +
   scale_color_viridis_d() +
   theme(legend.position = "bottom")


}
