% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes-step_box_cox.R, R/recipes-step_diff.R,
%   R/recipes-step_fourier.R, R/recipes-step_holiday_signature.R,
%   R/recipes-step_log_interval.R, R/recipes-step_slidify.R,
%   R/recipes-step_slidify_augment.R, R/recipes-step_smooth.R,
%   R/recipes-step_timeseries_signature.R, R/recipes-step_ts_clean.R,
%   R/recipes-step_ts_impute.R, R/recipes-step_ts_pad.R
\name{required_pkgs.step_box_cox}
\alias{required_pkgs.step_box_cox}
\alias{required_pkgs.step_diff}
\alias{required_pkgs.step_fourier}
\alias{required_pkgs.step_holiday_signature}
\alias{required_pkgs.step_log_interval}
\alias{required_pkgs.step_slidify}
\alias{required_pkgs.step_slidify_augment}
\alias{required_pkgs.step_smooth}
\alias{required_pkgs.step_timeseries_signature}
\alias{required_pkgs.step_ts_clean}
\alias{required_pkgs.step_ts_impute}
\alias{required_pkgs.step_ts_pad}
\title{S3 methods for tracking which additional packages are needed for steps.}
\usage{
required_pkgs.step_box_cox(x, ...)

required_pkgs.step_diff(x, ...)

required_pkgs.step_fourier(x, ...)

required_pkgs.step_holiday_signature(x, ...)

required_pkgs.step_log_interval(x, ...)

required_pkgs.step_slidify(x, ...)

required_pkgs.step_slidify_augment(x, ...)

required_pkgs.step_smooth(x, ...)

required_pkgs.step_timeseries_signature(x, ...)

required_pkgs.step_ts_clean(x, ...)

required_pkgs.step_ts_impute(x, ...)

required_pkgs.step_ts_pad(x, ...)
}
\arguments{
\item{x}{A recipe step}
}
\value{
A character vector
}
\description{
Recipe-adjacent packages always list themselves as a required package so that
the steps can function properly within parallel processing schemes.
}
\keyword{internal}
