% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes-step_fourier.R
\name{step_fourier_series}
\alias{step_fourier_series}
\alias{tidy.step_fourier_series}
\title{Fourier Features for Modeling Seasonality}
\usage{
step_fourier_series(
  recipe,
  ...,
  period,
  K,
  role = "predictor",
  trained = FALSE,
  columns = NULL,
  scale_factor = NULL,
  skip = FALSE,
  id = rand_id("fourier_series")
)

\method{tidy}{step_fourier_series}(x, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{A single column with class \code{Date} or
\code{POSIXct}. See \code{\link[recipes:selections]{recipes::selections()}} for more details.
For the \code{tidy} method, these are not currently used.}

\item{period}{The numeric period for the oscillation frequency.
See details for examples of \code{period} specification.}

\item{K}{The number of orders to include for each sine/cosine
fourier term. More orders increase the number of fourier terms and
therefore the variance of the fitted
model at the expense of bias. See details for examples of \code{K} specification.}

\item{role}{For model terms created by this step, what analysis
role should they be assigned?. By default, the function assumes
that the new variable columns created by the original variables
will be used as predictors in a model.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{columns}{A character string of variables that will be
used as inputs. This field is a placeholder and will be
populated once \code{\link[recipes:prep.recipe]{recipes::prep.recipe()}} is used.}

\item{scale_factor}{A factor for scaling the numeric index extracted
from the date or date-time feature. This is a placeholder and will be populated
once \code{\link[recipes:prep.recipe]{recipes::prep.recipe()}} is used.}

\item{skip}{A logical. Should the step be skipped when the recipe is
baked by bake.recipe()? While all operations are baked when prep.recipe()
is run, some operations may not be able to be conducted on new data
(e.g. processing the outcome variable(s)). Care should be taken when
using skip = TRUE as it may affect the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}

\item{x}{A \code{step_fourier_series} object.}
}
\value{
For \code{step_fourier_series}, an updated version of recipe with
the new step added to the sequence of existing steps (if any).
For the \code{tidy} method, a tibble with columns \code{terms}
(the selectors or variables selected), \code{value} (the feature
names).
}
\description{
\code{step_fourier_series} creates a a \emph{specification} of a recipe
step that will convert a Date or Date-time column into a Fourier
series
}
\details{
\strong{Date Variable}
Unlike other steps, \code{step_fourier_series} does \emph{not}
remove the original date variables. \code{\link[recipes:step_rm]{recipes::step_rm()}} can be
used for this purpose.

\strong{Period Specification}
The \code{period} argument is used to generate the distance between peaks
in the fourier sequence. The key is to line up the peaks with unique
seasonalities in the data.

For Daily Data, typical period specifications are:
\itemize{
\item Yearly frequency is 365
\item Quarterly frequency is 365 / 4 = 91.25
\item Monthly frequency is 365 / 12 = 30.42
}

\strong{K Specification}
The \code{K} argument specifies the maximum number of orders of Fourier terms.
Examples:
\itemize{
\item Specifying \code{period = 365} and \code{K = 1} will return a \code{cos_1_365} and \code{sin_1_365} sequence
\item Specifying \code{period = 365} and \code{K = 2} will return a \code{cos_1_365}, \code{cos_2_365}, \code{sin_1_365} and \code{sin_2_365}
sequence, which tends to increase the models ability to fit vs the \code{K = 1} specification
(at the expense of possibly overfitting).
}

\strong{Multiple values of \code{period} and \code{K}}
It's possible to specify multiple values of \code{period} in a single
step such as \code{step_fourier_series(period = c(91.25, 365), K = c(1, 1))}.
Note that the number of \code{K} elements must match the number of \code{period} elements.
}
\examples{
library(recipes)
library(tidyverse)
library(tidyquant)
library(timetk)

FB_tbl <- FANG \%>\% filter(symbol == "FB")

# Create a recipe object with a timeseries signature step
rec_obj <- recipe(adjusted ~ ., data = FB_tbl) \%>\%
    step_fourier_series(date, period = c(90.25, 365), K = c(1, 1))

# View the recipe object
rec_obj

# Prepare the recipe object
prep(rec_obj)

# Bake the recipe object - Adds the Fourier Series
bake(prep(rec_obj), FB_tbl)

# Tidy shows which features have been added during the 1st step
#  in this case, step 1 is the step_timeseries_signature step
tidy(rec_obj)
tidy(rec_obj, number = 1)

}
\seealso{
\code{\link[=step_timeseries_signature]{step_timeseries_signature()}} \code{\link[recipes:step_holiday]{recipes::step_holiday()}} \code{\link[recipes:step_date]{recipes::step_date()}}
\code{\link[recipes:step_rm]{recipes::step_rm()}}
\code{\link[recipes:recipe]{recipes::recipe()}} \code{\link[recipes:prep.recipe]{recipes::prep.recipe()}}
\code{\link[recipes:bake.recipe]{recipes::bake.recipe()}}
}
\concept{dates}
\concept{model_specification}
\concept{preprocessing}
\concept{variable_encodings}
\keyword{datagen}
