% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-summarise_by_time.R
\name{summarise_by_time}
\alias{summarise_by_time}
\alias{summarize_by_time}
\title{Summarise each group by time}
\usage{
summarise_by_time(
  .data,
  .date_var,
  ...,
  .by = "day",
  .type = c("floor", "ceiling", "round")
)

summarize_by_time(
  .data,
  .date_var,
  ...,
  .by = "day",
  .type = c("floor", "ceiling", "round")
)
}
\arguments{
\item{.data}{A \code{tbl} object or \code{data.frame}}

\item{.date_var}{A column of date or date-time (e.g. POSIXct) data class}

\item{...}{Name-value pairs of summary functions.
The name will be the name of the variable in the result.

The value can be:
\itemize{
\item A vector of length 1, e.g. \code{min(x)}, \code{n()}, or \code{sum(is.na(y))}.
\item A vector of length \code{n}, e.g. \code{quantile()}.
\item A data frame, to add multiple columns from a single expression.
}}

\item{.by}{A time unit to summarise by.
Time units are collapsed using \code{lubridate::floor_date()} or \code{lubridate::ceiling_date()}.

The value can be:
\itemize{
\item \code{second}
\item \code{minute}
\item \code{hour}
\item \code{day}
\item \code{week}
\item \code{month}
\item \code{bimonth}
\item \code{quarter}
\item \code{season}
\item \code{halfyear}
\item \code{year}
}

Arbitrary unique English abbreviations as in the \code{lubridate::period()} constructor are allowed.}

\item{.type}{One of "floor", "ceiling", or "round. Defaults to "floor". See \code{lubridate::round_date}.}
}
\value{
An object \emph{usually} of the same type as \code{.data}.
\itemize{
\item The rows come from the underlying \code{group_keys()}.
\item The columns are a combination of the grouping keys and the summary
expressions that you provide.
\item If \code{x} is grouped by more than one variable, the output will be another
\link{grouped_df} with the right-most group removed.
\item If \code{x} is grouped by one variable, or is not grouped, the output will
be a \link{tibble}.
\item Data frame attributes are \strong{not} preserved, because \code{summarise()}
fundamentally creates a new data frame.
}
}
\description{
\code{summarise_by_time()} is a time-based variant of the popular \code{dplyr::summarise()} function.
This time-based variant uses \code{.date_var} to specify a date or date-time column and \code{.by} to group the
calculation.

\code{summarise_by_time()} and \code{summarize_by_time()} are synonyms.
}
\section{Useful summary functions}{

\itemize{
\item Sum: \code{\link[=sum]{sum()}}
\item Center: \code{\link[=mean]{mean()}}, \code{\link[=median]{median()}}
\item Spread: \code{\link[=sd]{sd()}}, \code{\link[=var]{var()}}
\item Range: \code{\link[=min]{min()}}, \code{\link[=max]{max()}}
\item Count: \code{\link[dplyr:n]{dplyr::n()}}, \code{\link[dplyr:n_distinct]{dplyr::n_distinct()}}
\item Position: \code{\link[dplyr:first]{dplyr::first()}}, \code{\link[dplyr:last]{dplyr::last()}}, \code{\link[dplyr:nth]{dplyr::nth()}}
\item Correlation: \code{\link[=cor]{cor()}}, \code{\link[=cov]{cov()}}
}
}

\section{Methods}{

This function is a \strong{generic}, which means that packages can provide
implementations (methods) for other classes. See the documentation of
individual methods for extra arguments and differences in behaviour.
}

\examples{
# Libraries
library(tidyquant)
library(timetk)
library(dplyr)

# First adjusted price in each month
FANG \%>\%
    group_by(symbol) \%>\%
    summarise_by_time(
        date, .by = "month", # Setup for monthly aggregation
        adjusted  = FIRST(adjusted)
    )

# Last adjused price in each month (day is last day of month with ceiling option)
FANG \%>\%
    group_by(symbol) \%>\%
    summarise_by_time(
        .date_var  = date,
        .by        = "month",
        adjusted   = LAST(adjusted),
        .type      = "ceiling")

# Total Volume each year (.by is set to "year" now)
FANG \%>\%
    group_by(symbol) \%>\%
    summarise_by_time(
        .date_var  = date,
        .by        = "year",
        adjusted   = SUM(volume))


}
\concept{single table verbs}
