% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_augment.R
\name{sw_augment.default}
\alias{sw_augment.default}
\title{Default augment method}
\usage{
\method{sw_augment}{default}(x, ...)
}
\arguments{
\item{x}{an object to be tidied}

\item{...}{extra arguments passed to \code{broom::augment()}}
}
\value{
A tibble generated by \code{\link[broom:augment]{broom::augment()}}
}
\description{
By default, \code{sw_augment()} uses \code{\link[broom:augment]{broom::augment()}} to convert its output.
}
