% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_decompose.R
\name{sw_decompose}
\alias{sw_decompose}
\title{Coerces decomposed time-series objects to tibble format.}
\usage{
sw_decompose(model, index_rename = "index", ...)
}
\arguments{
\item{model}{A time-series object of class \code{stl}, \code{ets}, \code{decomposed.ts}.}

\item{index_rename}{Enables the index column to be renamed.}

\item{...}{Additional parameters passed to the \code{\link[tibble:as_tibble]{tibble::as_tibble()}} function.}
}
\value{
Returns a \code{tibble} object.
}
\description{
Coerces decomposed time-series objects to tibble format.
}
\details{
\code{sw_decompose} is designed
to decompose various time-series objects from Rob Hyndman's excellent \code{forecast} package
to \code{tibble} objects.
A regularized time index is always constructed. If no time index is
detected, a sequential index is returned as a default.
The index column name can be changed using the \code{index_rename} argument.
}
\examples{
library(forecast)
library(sweep)

# Decompose ETS model
USAccDeaths \%>\%
    ets() \%>\%
    sw_decompose()

# Decompose STL object
USAccDeaths \%>\%
    stl(s.window = 'periodic') \%>\%
    sw_decompose()


}
