% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_glance.R
\name{sw_glance.default}
\alias{sw_glance.default}
\title{Default glance method}
\usage{
\method{sw_glance}{default}(x, ...)
}
\arguments{
\item{x}{an object to be tidied}

\item{...}{extra arguments passed to \code{broom::glance()}}
}
\value{
A tibble generated by \code{\link[broom:glance]{broom::glance()}}
}
\description{
By default, \code{sw_glance()} uses \code{\link[broom:glance]{broom::glance()}} to convert its output.
}
