% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_sweep.R
\name{sw_sweep}
\alias{sw_sweep}
\title{Turn forecast objects into tibbles.}
\usage{
sw_sweep(x, .fitted = TRUE, index_rename = "index", ...)
}
\arguments{
\item{x}{A time-series forecast of class \code{forecast}.}

\item{.fitted}{Whether or not to return the fitted value in the results}

\item{index_rename}{Enables the index column to be renamed.}

\item{...}{Additional parameters passed to the \code{\link[tibble:as_tibble]{tibble::as_tibble()}} function.}
}
\value{
Returns a \code{tibble} object.
}
\description{
Turn forecast objects into tibbles.
}
\details{
\code{sw_sweep} is designed
to coerce \code{forecast} objects from the \code{forecast} package
to \code{tibble} objects. The returned object contains both the actual values
and the forecasted values including the point forecast and upper and lower
confidence intervals.
A regularized time index is always constructed. If no time index is
detected, a sequential index is returned as a default.
The index column name can be changed using the \code{index_rename} argument.
}
\examples{
library(forecast)
library(sweep)

# ETS forecasts
USAccDeaths \%>\%
    ets() \%>\%
    forecast(level = c(80, 95, 99)) \%>\%
    sw_sweep()


}
