% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_tbl.R
\name{sw_tbl}
\alias{sw_tbl}
\title{Coerce time-series objects to tibble.}
\usage{
sw_tbl(data, preserve_index = TRUE, index_rename = "index", ...)
}
\arguments{
\item{data}{A time-series object.}

\item{preserve_index}{Attempts to preserve a time series index. Default is \code{TRUE}.}

\item{index_rename}{Enables the index column to be renamed.}

\item{...}{Additional parameters passed to the \code{\link[tibble:as_tibble]{tibble::as_tibble()}} function.}
}
\value{
Returns a \code{tibble} object.
}
\description{
Coerce time-series objects to tibble.
}
\details{
\code{sw_tbl} is designed
to coerce time series objects (e.g. \code{xts}, \code{zoo}, \code{ts}, \code{timeSeries}, etc)
to \code{tibble} objects. The main advantage is that the function attempts to keep the
date / date-time information from the underlying time-series object.
When \code{preserve_index = TRUE} is specified, a new column,
\code{index}, is created during object coercion, and the function attempts to preserve
the date or date-time information. The date / date-time column name
can be changed using the \code{index_rename} argument.
}
\examples{
### ts to tibble: Comparison between as.data.frame() and sw_tbl()
data_ts <- ts(1:10, start = c(2015,1), freq = 4)

# No index
as.data.frame(data_ts)

# Regularized numeric index starting in 2015
sw_tbl(data_ts)


### zooreg to tibble: Comparison between as.data.frame() and sw_tbl()
data_zooreg <- zoo::zooreg(1:8, start = zoo::yearqtr(2000), frequency = 4)

# Dates are character class stored in row names
as.data.frame(data_zooreg)

# Dates are appropriate zoo yearqtr class within the data frame
sw_tbl(data_zooreg)


### zoo to tibble: Comparison between as.data.frame() and sw_tbl()
data_zoo <- zoo::zoo(1:12, zoo::yearmon(2016 + seq(0, 11)/12))

# Dates are character class stored in row names
as.data.frame(data_zoo)

# Dates are appropriate zoo yearmon class within the data frame
sw_tbl(data_zoo)


### xts to tibble: Comparison between as.data.frame() and sw_tbl()
data_xts <- xts::as.xts(
    x        = matrix(c(1:5 * 10, 5:1 * rnorm(1)), ncol = 2),
    order.by = seq.Date(from       = as.Date("2010-01-01"),
                        by         = 1,
                        length.out = 5)
    )
colnames(data_xts) <- c("x", "y")

# Dates are character class stored in row names
as.data.frame(data_xts)

# Dates are appropriate date class and within the data frame
sw_tbl(data_xts)


}
\seealso{
\code{\link[=sw_xts]{sw_xts()}}, \code{\link[=sw_zoo]{sw_zoo()}}, \code{\link[=sw_zooreg]{sw_zooreg()}}, \code{\link[=sw_ts]{sw_ts()}}
}
