% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_tidy.R
\name{sw_tidy}
\alias{sw_tidy}
\title{Tidy the result of a time-series model into a summary tibble}
\usage{
sw_tidy(x, ...)
}
\arguments{
\item{x}{An object to be converted into a tibble ("tidy" data.frame)}

\item{...}{extra arguments}
}
\value{
a tibble
}
\description{
Tidy the result of a time-series model into a summary tibble
}
\details{
\code{sw_tidy()} methods always returns a "tidy" tibble with model coeffc.
The single row includes summary statistics relevent to the model accuracy,
which can be used to assess model fit and quality. The benefit of \code{sw_glance}
is that it has methods for various time-series model classes such as
\code{HoltWinters}, \code{ets}, \code{Arima}, etc.

For non-time series, \code{sw_glance()} defaults to \code{broom::glance()}.
The only difference is that the return is a tibble.
The output of \code{sw_tidy()} is always a tibble with disposable row names. It is
therefore suited for further manipulation by packages like dplyr and
ggplot2. The \code{sw_tidy()} function is designed to work well with time-series
models (supplied via \code{x}). The function provides additional "tidy" functionality
to the \code{broom::tidy()} function as follows:
\enumerate{
\item Model object classes such as \code{ets}, \code{Arima}, \code{bats}, \code{stl},
and others from Rob Hyndman's excellent \code{forecast} package.
\item Model object classes such as \code{HoltWinters} and \code{Arima} from the \code{stats}
package.
}
}
\examples{
library(forecast)
WWWusage \%>\%
    auto.arima() \%>\%
    sw_tidy(conf.int = TRUE)

}
\seealso{
\code{\link[=tidy]{tidy()}}
}
