% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_ts.R
\name{sw_ts}
\alias{sw_ts}
\alias{sw_ts_}
\title{Coerce time series objects and tibbles with date/date-time columns to ts.}
\usage{
sw_ts(data, select = NULL, start = 1, end = numeric(), frequency = 1,
  deltat = 1, ts.eps = getOption("ts.eps"))

sw_ts_(data, select = NULL, start = 1, end = numeric(), frequency = 1,
  deltat = 1, ts.eps = getOption("ts.eps"))
}
\arguments{
\item{data}{A time-based tibble or time-series object.}

\item{select}{\strong{Applicable to tibbles and data frames only}.
The column or set of columns to be coerced to \code{ts} class.}

\item{start}{the time of the first observation.  Either a single
    number or a vector of two integers, which specify a natural time
    unit and a (1-based) number of samples into the time unit.  See
    the examples for the use of the second form.}

\item{end}{the time of the last observation, specified in the same way
    as \code{start}.}

\item{frequency}{the number of observations per unit of time.}

\item{deltat}{the fraction of the sampling period between successive
    observations; e.g., 1/12 for monthly data.  Only one of
    \code{frequency} or \code{deltat} should be provided.}

\item{ts.eps}{time series comparison tolerance.  Frequencies are
    considered equal if their absolute difference is less than
    \code{ts.eps}.}
}
\value{
Returns a \code{ts} object.
}
\description{
Coerce time series objects and tibbles with date/date-time columns to ts.
}
\details{
\code{sw_ts()} is a wrapper for \code{stats::ts()} that is designed
to coerce \code{tibble} objects that have a "time-base" (meaning the values vary with time)
to \code{ts} class objects. There are two main advantages:
\enumerate{
\item Non-numeric columns get removed instead of being populated by NA's.
\item The returned \code{ts} object retains an index (and various other attributes) if detected.
}

The \code{select} argument is used to select subsets
of columns from the incoming data.frame.
Only columns containing numeric data are coerced. \emph{At a minimum, a \code{frequency}
and a \code{start} should be specified.}

For non-data.frame object classes (e.g. \code{xts}, \code{zoo}, \code{timeSeries}, etc) the objects are coerced
using \code{stats::ts()}.

\code{sw_ts_} is a nonstandard evaluation method.
}
\examples{
### tibble to ts: Comparison between sw_ts() and stats::ts()
data_tbl <- tibble::tibble(
    date = seq.Date(as.Date("2016-01-01"), by = 1, length.out = 5),
    x    = rep("chr values", 5),
    y    = cumsum(1:5),
    z    = cumsum(11:15) * rnorm(1))

# as.ts: Character columns introduce NA's; Result does not retain index
stats::ts(data_tbl[,-1], start = 2016)

# sw_ts: Only numeric columns get coerced; Result retains index in numeric format
sw_ts(data_tbl, start = 2016)


### Using select
sw_ts(data_tbl, select = y)


### NSE: Enables programming
select   <- "y"
sw_ts_(data_tbl, select = select)

}
\seealso{
\code{\link[=sw_tbl]{sw_tbl()}}, \code{\link[=sw_xts]{sw_xts()}}, \code{\link[=sw_zoo]{sw_zoo()}}, \code{\link[=sw_zooreg]{sw_zooreg()}}
}
