% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_unroll_index.R
\name{sw_unroll_index}
\alias{sw_unroll_index}
\title{Extract an index of date or datetime from time series}
\usage{
sw_unroll_index(data)
}
\arguments{
\item{data}{A time-based tibble, time-series object, time-series model,
or \code{forecast} object.}
}
\value{
Returns a vector of date or date times
}
\description{
Extract an index of date or datetime from time series
}
\details{
\code{sw_unroll_index} is used to extract the date or datetime index from various
time series data types.
The method is particularly useful for retrieving date or datetime index information
from \code{ts} objects which traditionally lose this information during coercion
from time-based tibble (\code{tbl}), \code{xts}, and \code{zoo} data types that contain
and index.
The method can additionally be used on \code{forecast} objects and a number of
objects generated by modeling functions such as \code{Arima}, \code{ets}, and \code{HoltWinters}
classes.

\strong{Important Note}: To gain the benefit of \code{sw_unroll_index()} the time series
must have an index. This is particularly important for \code{ts} objects, which
by default do not contain an index and therefore must be coerced from time-based
objects such as \code{tbl}, \code{xts}, or \code{zoo} using the \code{sw_ts()} function.
Refer to \code{\link[=sw_ts]{sw_ts()}} for creating persistent date / datetime index
during coercion to \code{ts}.
}
\examples{
library(tidyverse)
library(sweep)

# Create time-based tibble
data_tbl <- tibble::tibble(
    date = seq.Date(from = as.Date("2000-01-01"), by = 1, length.out = 5),
    x    = rnorm(5) * 10,
    y    = 5:1
)
sw_unroll_index(data_tbl) # Returns time-based index vector

# Coerce to ts using sw_ts()
data_ts <- sw_ts(data_tbl)
sw_unroll_index(data_ts)  # Returns original time-based index vector

# Coerce to tbl with original, non-regularized time index
sw_tbl(data_ts, .unroll = TRUE)

# Coerce to xts
data_ts \%>\%
    sw_xts(order.by = sw_unroll_index(.))

# Coerce to xts
data_ts \%>\%
    sw_zoo(order.by = sw_unroll_index(.))


}
\seealso{
\code{\link[=sw_ts]{sw_ts()}}, \code{\link[=sw_tbl]{sw_tbl()}}, \code{\link[=sw_xts]{sw_xts()}}, \code{\link[=sw_zoo]{sw_zoo()}}, \code{\link[=sw_zooreg]{sw_zooreg()}}
}
