% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-taylor_30_min.R
\docType{data}
\name{taylor_30_min}
\alias{taylor_30_min}
\title{Half-hourly electricity demand}
\format{
A tibble: 4,032 x 2
\itemize{
\item \code{date}: A date-time variable in 30-minute increments
\item \code{value}: Electricity demand in Megawatts
}
}
\source{
James W Taylor
}
\usage{
taylor_30_min
}
\description{
Half-hourly electricity demand in England and Wales from Monday 5 June 2000
to Sunday 27 August 2000. Discussed in Taylor (2003).
}
\examples{
taylor_30_min

}
\references{
Taylor, J.W. (2003) Short-term electricity demand forecasting
using double seasonal exponential smoothing. \emph{Journal of the
Operational Research Society}, \bold{54}, 799-805.
}
\keyword{datasets}
