% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-forecast.R
\name{tbats_string}
\alias{tbats_string}
\title{Print the TBATS model parameters}
\usage{
tbats_string(object)
}
\arguments{
\item{object}{An object of class bats or tbats}
}
\description{
Refer to forecast:::makeTextTBATS.
\href{https://github.com/robjhyndman/forecast/blob/master/R/bats.R}{forecast bats.R}
}
