% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers_nested_df.R
\name{tidiers_nested_df}
\alias{tidiers_nested_df}
\alias{sw_tidy.data.frame}
\alias{sw_glance.data.frame}
\alias{sw_augment.data.frame}
\alias{sw_tidy_decomp.data.frame}
\alias{sw_sweep.data.frame}
\title{Tidying methods that work with nested data frames (list-columns)}
\usage{
\method{sw_tidy}{data.frame}(data, x, ...)

\method{sw_glance}{data.frame}(data, x, ...)

\method{sw_augment}{data.frame}(data, x, ...)

\method{sw_tidy_decomp}{data.frame}(data, x, ...)

\method{sw_sweep}{data.frame}(data, x, ...)
}
\arguments{
\item{data}{a nested tibble}

\item{x}{the column name of the column containing the models to
be tidied. For tidy, augment, and glance it should be the bare name; for
_ methods it should be quoted.}

\item{...}{additional arguments to pass on to the respective tidying method}
}
\value{
A \code{grouped_df}, where the non-list columns of the
original are used as grouping columns alongside the tidied outputs.
}
\description{
Tidying methods that work with nested data frames (list-columns)
}
\details{
These methods are designed to expedite tidying on data frames that are
nested meaning the columns contain lists that with contents
to be tidied by one of the \code{sweep} tidying functions.
}
\examples{
library(tidyverse)
library(sweep)

regressions <- mtcars \%>\%
    group_by(cyl) \%>\%
    do(mod = lm(mpg ~ wt, .))

regressions

regressions \%>\% sw_tidy(mod)
regressions \%>\% sw_glance(mod)
regressions \%>\% sw_augment(mod)

}
