% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers_StructTS.R
\name{tidiers_StructTS}
\alias{tidiers_StructTS}
\alias{sw_tidy.StructTS}
\alias{sw_glance.StructTS}
\alias{sw_augment.StructTS}
\title{Tidying methods for StructTS (Error, Trend, Seasonal) / exponential smoothing
modeling of time series}
\usage{
\method{sw_tidy}{StructTS}(x, ...)

\method{sw_glance}{StructTS}(x, ...)

\method{sw_augment}{StructTS}(x, data = NULL, index_rename = "index", ...)
}
\arguments{
\item{x}{An object of class "StructTS"}

\item{...}{Additional parameters (not used)}

\item{data}{Used with \code{sw_augment} only.
\code{NULL} by default which simply returns augmented columns only.
User can supply the original data, which returns the data + augmented columns.}

\item{index_rename}{Used with \code{sw_augment} only.
A string representing the name of the index generated.}
}
\value{
\strong{\code{sw_tidy()}} returns one row for each model parameter,
with two columns:
\itemize{
\item \code{term}: The model parameters
\item \code{estimate}: The estimated parameter value
}

\strong{\code{sw_glance()}} returns one row with the columns
\itemize{
\item \code{model.desc}: A description of the model including the
three integer components (p, d, q) are the AR order,
the degree of differencing, and the MA order.
\item \code{sigma}: The square root of the estimated residual variance
\item \code{logLik}: The data's log-likelihood under the model
\item \code{AIC}: The Akaike Information Criterion
\item \code{BIC}: The Bayesian Information Criterion
\item \code{ME}: Mean error
\item \code{RMSE}: Root mean squared error
\item \code{MAE}: Mean absolute error
\item \code{MPE}: Mean percentage error
\item \code{MAPE}: Mean absolute percentage error
\item \code{MASE}: Mean absolute scaled error
\item \code{ACF1}: Autocorrelation of errors at lag 1
}

\strong{\code{sw_augment()}} returns a tibble with the following time series attributes:
\itemize{
\item \code{index}: An index is either attempted to be extracted from the model or
a sequential index is created for plotting purposes
\item \code{.actual}: The original time series
\item \code{.fitted}: The fitted values from the model
\item \code{.resid}: The residual values from the model
}
}
\description{
These methods tidy the coefficients of StructTS models of univariate time
series.
}
\examples{
library(forecast)
library(sweep)

fit_StructTS <- WWWusage \%>\%
    StructTS()

sw_tidy(fit_StructTS)
sw_glance(fit_StructTS)
sw_augment(fit_StructTS)

}
\seealso{
\code{\link[=StructTS]{StructTS()}}
}
