% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timekit-package.R
\docType{package}
\name{timekit}
\alias{timekit}
\alias{timekit-package}
\title{timekit: a toolkit for time series}
\description{
The \code{timekit} package combines a collection of coercion tools for
time series analysis.
}
\details{
The \code{timekit} package has two main benefits:
\enumerate{
\item It simplifies the coercion process between time-based tibbles (\code{tbl})
and the major time series data types \code{xts}, \code{zoo}, \code{zooreg}, and \code{ts}
\item It maximizes time-based data retention during coercion to regularized time series
}

To learn more about \code{timekit}, start with the vignettes:
\code{browseVignettes(package = "timekit")}
}
