% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetk-package.R
\docType{package}
\name{timetk-package}
\alias{timetk}
\alias{timetk-package}
\title{timetk: Time Series Analysis in the Tidyverse}
\description{
The \code{timetk} package combines a collection of coercion tools for
time series analysis.
}
\details{
The \code{timetk} package has several benefits:
\enumerate{
\item Visualizing Time Series
\item Wrangling Time Series.
\item Preprocessing and Feature Engineering.
}

To learn more about \code{timetk}, start with the documentation:
\url{https://business-science.github.io/timetk/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/business-science/timetk}
  \item \url{https://business-science.github.io/timetk/}
  \item Report bugs at \url{https://github.com/business-science/timetk/issues}
}

}
\author{
\strong{Maintainer}: Matt Dancho \email{mdancho@business-science.io}

Authors:
\itemize{
  \item Davis Vaughan \email{dvaughan@business-science.io}
}

}
