% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetk-package.R
\docType{package}
\name{timetk}
\alias{timetk}
\alias{timetk-package}
\title{timetk: a toolkit for time series}
\description{
The \code{timetk} package combines a collection of coercion tools for
time series analysis.
}
\details{
The \code{timetk} package has several benefits:
\enumerate{
\item Index extraction: get the time series index from any time series object.
\item Understand time series: create a signature and summary from a time series index.
\item Build future time series: create a future time series from an index.
\item Coerce between time-based tibbles (tbl) and the major time series data types xts, zoo, zooreg, and ts: Simplifies coercion and maximizes time-based data retention during coercion to regularized time series (e.g. ts).
}

To learn more about \code{timetk}, start with the vignettes:
\code{browseVignettes(package = "timetk")}
}
