% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-tk_augment_differences.R
\name{tk_augment_differences}
\alias{tk_augment_differences}
\title{Augment columns with differenced data}
\usage{
tk_augment_differences(
  .data,
  .column,
  .lags = 1,
  .difference = 1,
  .log = FALSE,
  .names = paste0("diff_", .lags)
)
}
\arguments{
\item{.data}{A tibble.}

\item{.column}{A column to have a difference transformation applied}

\item{.lags}{One or more lags for the difference(s)}

\item{.difference}{The number of differences to apply.}

\item{.log}{If TRUE, applies log-differences.}

\item{.names}{A vector of names for the new columns. Must be of same length as \code{.lags}.}
}
\value{
Returns a \code{tibble} object describing the timeseries.
}
\description{
A handy function for adding multiple lagged difference values to a data frame.
Works group-wise.
}
\details{
\strong{Benefits}

This is a scalable function that is:
\itemize{
\item Designed to work with grouped data using \code{dplyr::group_by()}
\item Add multiple differences by adding a sequence of differences using
the \code{.lags} argument (e.g. \code{lags = 1:20})
}
}
\examples{
library(tidyverse)
library(timetk)

m4_monthly \%>\%
    group_by(id) \%>\%
    tk_augment_differences(value, .lags = 1:20)

}
\seealso{
\itemize{
\item \code{\link[=diff_vec]{diff_vec()}} - Underlying function that powers \code{tk_augment_differences()}
}
}
