% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk_augment_holiday.R
\name{tk_augment_holiday}
\alias{tk_augment_holiday}
\alias{tk_augment_holiday_signature}
\title{Augment the holiday signature to the data}
\usage{
tk_augment_holiday_signature(
  .data,
  .holiday_pattern = ".",
  .locale_set = c("all", "none", "World", "US", "CA", "GB", "FR", "IT", "JP", "CH",
    "DE"),
  .exchange_set = c("all", "none", "NYSE", "LONDON", "NERC", "TSX", "ZURICH")
)
}
\arguments{
\item{.data}{A time-based tibble or time-series object.}

\item{.holiday_pattern}{A regular expression pattern to search the "Holiday Set".}

\item{.locale_set}{Return binary holidays based on locale.
One of: "all", "none", "World", "US", "CA", "GB", "FR", "IT", "JP", "CH", "DE".}

\item{.exchange_set}{Return binary holidays based on Stock Exchange Calendars.
One of: "all", "none", "NYSE", "LONDON", "NERC", "TSX", "ZURICH".}
}
\value{
Returns a \code{tibble} object describing the holiday timeseries.
}
\description{
Augment the holiday signature to the data
}
\details{
\code{tk_augment_holiday_signature} adds the holiday signature
features. See \code{\link[=tk_get_holiday_signature]{tk_get_holiday_signature()}} (powers the augment function)
for a full description and examples for how to use.

\strong{1. Individual Holidays}

These are \strong{single holiday features} that can be filtered using a pattern.
This helps in identifying which holidays are important to a machine learning model.
This can be useful for example in \strong{e-commerce initiatives}
(e.g. sales during Christmas and Thanskgiving).

\strong{2. Locale-Based Summary Sets}

Locale-based holdiay sets are useful for \strong{e-commerce initiatives}
(e.g. sales during Christmas and Thanskgiving). Filter on a locale to
identify all holidays in that locale.

\strong{3. Stock Exchange Calendar Summary Sets}

Exchange-based holdiay sets are useful for identifying \strong{non-working days.}
Filter on an index to identify all holidays that are commonly non-working.
}
\examples{
library(dplyr)
library(tidyquant)
library(timetk)

FANG \%>\%
    filter(symbol == "FB") \%>\%
    tk_augment_holiday_signature()

}
\seealso{
\itemize{
\item \code{\link[=tk_get_holiday_signature]{tk_get_holiday_signature()}} - Underlying function that powers holiday feature generation
}
}
