% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-tk_augment_lags.R
\name{tk_augment_lags}
\alias{tk_augment_lags}
\title{Augment columns with lagged data}
\usage{
tk_augment_lags(.data, .column, .lags = 1, .names = paste0("lag_", .lags))
}
\arguments{
\item{.data}{A tibble.}

\item{.column}{A column to have a difference transformation applied}

\item{.lags}{One or more lags for the difference(s)}

\item{.names}{A vector of names for the new columns. Must be of same length as \code{.lags}.}
}
\value{
Returns a \code{tibble} object describing the timeseries.
}
\description{
A handy function for adding multiple lagged columns to a data frame.
Works group-wise.
}
\details{
\strong{Benefits}

This is a scalable function that is:
\itemize{
\item Designed to work with grouped data using \code{dplyr::group_by()}
\item Add multiple lags by adding a sequence of lags using
the \code{.lags} argument (e.g. \code{.lags = 1:20})
}
}
\examples{
library(tidyverse)
library(timetk)

m4_monthly \%>\%
    group_by(id) \%>\%
    tk_augment_lags(value, .lags = 1:20)

}
\seealso{
Augment Operations:
\itemize{
\item \code{\link[=tk_augment_timeseries_signature]{tk_augment_timeseries_signature()}} - Group-wise augmentation of timestamp features
\item \code{\link[=tk_augment_holiday_signature]{tk_augment_holiday_signature()}} - Group-wise augmentation of holiday features
\item \code{\link[=tk_augment_roll_apply]{tk_augment_roll_apply()}} - Group-wise augmentation of rolling functions
\item \code{\link[=tk_augment_lags]{tk_augment_lags()}} - Group-wise augmentation of lagged data
\item \code{\link[=tk_augment_differences]{tk_augment_differences()}} - Group-wise augmentation of differenced data
}

Underlying Function:
\itemize{
\item \code{\link[=lag_vec]{lag_vec()}} - Underlying function that powers \code{tk_augment_lags()}
}
}
