% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-tk_augment_timeseries.R
\name{tk_augment_timeseries}
\alias{tk_augment_timeseries}
\alias{tk_augment_timeseries_signature}
\title{Augment the time series signature to the data}
\usage{
tk_augment_timeseries_signature(.data)
}
\arguments{
\item{.data}{A time-based tibble or time-series object.}
}
\value{
Returns a \code{tibble} object describing the timeseries.
}
\description{
Augment the time series signature to the data
}
\details{
\code{tk_augment_timeseries_signature} adds the time series signature
features including
numeric value, differences,
year, month, day, day of week, day of month,
day of year, hour, minute, second
to the input data.
}
\examples{
library(dplyr)
library(tidyquant)
library(timetk)

FANG \%>\%
    filter(symbol == "FB") \%>\%
    tk_augment_timeseries_signature()

}
\seealso{
Augment Operations:
\itemize{
\item \code{\link[=tk_augment_timeseries_signature]{tk_augment_timeseries_signature()}} - Group-wise augmentation of timestamp features
\item \code{\link[=tk_augment_holiday_signature]{tk_augment_holiday_signature()}} - Group-wise augmentation of holiday features
\item \code{\link[=tk_augment_roll_apply]{tk_augment_roll_apply()}} - Group-wise augmentation of rolling functions
\item \code{\link[=tk_augment_lags]{tk_augment_lags()}} - Group-wise augmentation of lagged data
\item \code{\link[=tk_augment_differences]{tk_augment_differences()}} - Group-wise augmentation of differenced data
}

Underlying Function:
\itemize{
\item \code{\link[=tk_get_timeseries_signature]{tk_get_timeseries_signature()}} - Returns timeseries features from an index
}
}
