% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk_get_holiday_signature.R
\name{tk_get_holiday}
\alias{tk_get_holiday}
\alias{tk_get_holiday_signature}
\alias{tk_get_holidays_by_year}
\title{Get holiday features from a time-series index}
\usage{
tk_get_holiday_signature(
  idx,
  holiday_pattern = ".",
  locale_set = c("all", "none", "World", "US", "CA", "GB", "FR", "IT", "JP", "CH",
    "DE"),
  exchange_set = c("all", "none", "NYSE", "LONDON", "NERC", "TSX", "ZURICH")
)

tk_get_holidays_by_year(years = year(today()))
}
\arguments{
\item{idx}{A time-series index that is a vector of dates or datetimes.}

\item{holiday_pattern}{A regular expression pattern to search the "Holiday Set".}

\item{locale_set}{Return binary holidays based on locale.
One of: "all", "none", "World", "US", "CA", "GB", "FR", "IT", "JP", "CH", "DE".}

\item{exchange_set}{Return binary holidays based on Stock Exchange Calendars.
One of: "all", "none", "NYSE", "LONDON", "NERC", "TSX", "ZURICH".}

\item{years}{One or more years to collect holidays for.}
}
\value{
Returns a \code{tibble} object describing the timeseries holidays.
}
\description{
Get holiday features from a time-series index
}
\details{
TODO

3 Sets in 1:

\strong{1. Individual Holidays}

\strong{2. Locale-Based Summary Sets}

\strong{3. Stock Exchange Calendar Summary Sets}
}
\examples{
library(dplyr)
library(tidyquant)
library(timetk)

# Works with time-based tibbles
idx <- tk_make_index_sequence("2017-01-01", "2017-12-31", by = "day")

# --- BASIC USAGE ----

tk_get_holiday_signature(idx)

# ---- FILTERING WITH PATTERNS & SETS ----

# List available holidays - see patterns
tk_get_holidays_by_year(2020)

# Filter using holiday patterns
# - Get Christmas and Thanksgiving Features in US
tk_get_holiday_signature(
    idx,
    holiday_pattern = "(US_Christmas)|(US_Thanks)",
    locale_set      = "none",
    exchange_set    = "none")

# Filter with locale sets
tk_get_holiday_signature(
    idx,
    holiday_pattern = "$^", # Matches nothing on purpose
    locale_set      = "US",
    exchange_set    = "none")

# Filter with exchange sets
tk_get_holiday_signature(
    idx,
    holiday_pattern = "$^", # Matches nothing on purpose
    locale_set      = "none",
    exchange_set    = "NYSE")

}
\seealso{
TODO
}
