% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dates.R
\name{tk_get_timeseries_unit_frequency}
\alias{tk_get_timeseries_unit_frequency}
\title{Get the timeseries unit frequency for the primary time scales}
\usage{
tk_get_timeseries_unit_frequency()
}
\value{
\code{tk_get_timeseries_unit_frequency} returns a tibble containing
the timeseries frequencies in seconds for the primary time scales including
"sec", "min", "hour", "day", "week", "month", "quarter", and "year".
}
\description{
Get the timeseries unit frequency for the primary time scales
}
\examples{
tk_get_timeseries_unit_frequency()


}
