% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-tk_lag_diagnostics.R
\name{tk_lag_diagnostics}
\alias{tk_lag_diagnostics}
\title{Get ACF, PACF, and CCF in 1 Data Frame}
\usage{
tk_lag_diagnostics(.data, .value, ..., .lags = 0:60)
}
\arguments{
\item{.data}{A data frame or tibble with numeric features (values) in descending
chronological order}

\item{.value}{A numeric column with a value to have ACF and PACF calculations
performed.}

\item{...}{Additional features to perform Lag Cross Correlations (CCFs)
versus the \code{.value}. Useful for evaluating external lagged regressors.}

\item{.lags}{A seqence of one or more lags to evaluate.}
}
\description{
The \code{tk_lag_diagnostics()} function provides a simple interface to
detect Autocorrelation, Partial ACF, and Cross Correlation of Lagged
Predictors in one \code{tibble}. This function powers the \code{\link[=plot_lag_diagnostics]{plot_lag_diagnostics()}}
visualization.
}
\details{
\strong{Simplified ACF, PACF, & CCF}

We are often interested in all 3 of these functions. Why not get all 3 at once?
Now you can!
\itemize{
\item \strong{ACF} - Autocorrelation between a target variable and lagged versions of itself
\item \strong{PACF} - Partial Autocorrelation removes the dependence of lags on
other lags highlighting key seasonalities.
\item \strong{CCF} - Shows how lagged predictors can be used for prediction of a target
variable.
}

\strong{Works with Grouped Data Frames}

The \code{tk_lag_diagnostics()} works with \code{grouped_df}'s, meaning you can
group your time series by one or more categorical columns with \code{group_by()}
and then apply \code{tk_lag_diagnostics()} to return group-wise lag diagnostics.
}
\examples{
library(tidyverse)
library(tidyquant)
library(timetk)

# ACF, PACF, & CCF in 1 Data Frame
# - Get ACF & PACF for target (adjusted)
# - Get CCF between adjusted and volume and close
FANG \%>\%
    filter(symbol == "FB") \%>\%
    tk_lag_diagnostics(.value = adjusted, volume, close, .lags = 0:500)

# Do the same thing for groups with group_by()
FANG \%>\%
    group_by(symbol) \%>\%
    tk_lag_diagnostics(.value = adjusted, volume, close, .lags = 0:500)


}
\seealso{
\itemize{
\item \strong{Visualizing ACF, PACF, & CCF:} \code{\link[=plot_lag_diagnostics]{plot_lag_diagnostics()}}
}
}
