% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune-tk_tune_parameters.R
\name{tk_parameter_ranking}
\alias{tk_parameter_ranking}
\title{Rank Hyperparameter Tuning Results}
\usage{
tk_parameter_ranking(.data, .metric, .max_failure_rate = 1)
}
\arguments{
\item{.data}{A \code{tibble} of class "tune_results"}

\item{.metric}{Select a metric to use for tuning performance investigation.}

\item{.max_failure_rate}{Range (0,1). Use to \code{.max_failure_rate} to filter models below an acceptable failure threshold.}
}
\value{
A \code{tibble} or \code{data.frame} with ranked tuning parameters.
}
\description{
\code{tk_parameter_ranking()} returns the tuning performance results
from functions like \code{tune::tune_grid()}. If possible, adds parameter ranking
results based on metric performance, model failure rates, and standard error (variability).
}
\details{
\strong{Metric Ranking (Model Accuracy, Y-Axis)}

The model with the lowest (best) rank is that with the lowest mean error. Y-Axis is the Metric value.
\itemize{
\item \strong{Metric Calculation} - Refer to the appropriate metric from the \code{yardstick} R package.
\item \strong{Metric Ranking} - Sort Best to Worst mean metric. Rank 1 to N.
}

\strong{Failure Rate Ranking (Robustness to New Data, Color)}

Models with lower failure rates are more robust to new data. Color is the Failure Rate.
Failure rate rank is a score based on the proportion of models that failed during tuning.
\itemize{
\item \strong{Failure Rate Calculation:} \emph{Failure Rate = n / No. of Resample Slices}
\item \strong{Failure Rate Ranking:} Use Min Ranking (\code{dplyr::min_rank()}) on the Failure Rate Calculation.
}

\strong{Standard Error Ranking (Model Variability, Size)}

Models with lower standard error are more consistent (less variability). Size of the point is
the standard error ranking.
\itemize{
\item \strong{Standard Error Calculation} - See \code{base::sd()}.
\item \strong{Standard Error Ranking} - Sort Best to Worst standard error. Rank 1 to N.
}
}
\examples{
library(dplyr)
library(tune)
library(timetk)

# Output of tune::tune_grid()
arima_workflow_tuned

# PARAMETER RANKING ----
arima_workflow_tuned \%>\%
    tk_parameter_ranking(.max_failure_rate = 1)

# PARAMETER SELECTION ----
arima_workflow_tuned \%>\%
    tk_parameter_ranking(.max_failure_rate = 1) \%>\%
    tk_parameter_select_by_row(.n = 3)

}
