% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune-tk_tune_performance.R
\name{tk_tune_select_parameters}
\alias{tk_tune_select_parameters}
\title{Select Hyperparameters}
\usage{
tk_tune_select_parameters(.data, .n = 1)
}
\arguments{
\item{.data}{A \code{tibble} of class "tune_results"}

\item{.n}{Row to select from the model ranking}
}
\value{
A \code{tibble} or \code{data.frame} with parameter values.
}
\description{
\code{tk_tune_select_parameters()} provides a wrapper to \code{tune::select_best()} that
makes it easy to select the n-th model.
}
\examples{
library(dplyr)
library(tune)
library(timetk)

arima_workflow_tuned \%>\%
    tk_tune_rank_parameters(.max_failure_rate = 1) \%>\%
    tk_tune_select_parameters(.n = 3)

}
