% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-ts_impute.R
\name{ts_impute_vec}
\alias{ts_impute_vec}
\title{Missing Value Imputation for Time Series}
\usage{
ts_impute_vec(x, period = 1, lambda = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{period}{A seasonal period to use during the transformation. If \code{period = 1},
linear interpolation is performed. If \code{period > 1}, a robust STL decomposition is
first performed and a linear interpolation is applied to the seasonally adjusted data.}

\item{lambda}{A box cox transformation parameter. If set to \code{"auto"}, performs
automated lambda selection.}
}
\value{
A \code{numeric} vector with the missing values imputed.
}
\description{
This is mainly a wrapper for the Seasonally Adjusted Missing Value using Linear Interpolation function,
\code{na.interp()}, from the \code{forecast} R package. The \code{ts_impute_vec()} function includes arguments for applying
seasonality to numeric vector (non-\code{ts}) via the \code{period} argument.
}
\details{
\strong{Imputation using Linear Interpolation}

Three circumstances cause strictly linear interpolation:
\enumerate{
\item \strong{Period is 1:} With \code{period = 1}, a seasonality cannot be interpreted and therefore linear is used.
\item \strong{Number of Non-Missing Values is less than 2-Periods}: Insufficient values exist to detect seasonality.
\item \strong{Number of Total Values is less than 3-Periods}: Insufficient values exist to detect seasonality.
}

\strong{Seasonal Imputation using Linear Interpolation}

For seasonal series with \code{period > 1}, a robust Seasonal Trend Loess (STL) decomposition is first computed.
Then a linear interpolation is applied to the seasonally adjusted data, and
the seasonal component is added back.

\strong{Box Cox Transformation}

In many circumstances, a Box Cox transformation can help. Especially if the series is multiplicative
meaning the variance grows exponentially. A Box Cox transformation can be automated by setting \code{lambda = "auto"}
or can be specified by setting \verb{lambda = numeric value}.
}
\examples{
library(dplyr)


# --- VECTOR ----

values <- c(1,2,3, 4*2, 5,6,7, NA, 9,10,11, 12*2)
values

# Linear interpolation
ts_impute_vec(values, period = 1, lambda = NULL)

# Seasonal Interpolation: set period = 4
ts_impute_vec(values, period = 4, lambda = NULL)

# Seasonal Interpolation with Box Cox Transformation (internal)
ts_impute_vec(values, period = 4, lambda = "auto")


}
\references{
\itemize{
\item \href{https://github.com/robjhyndman/forecast}{Forecast R Package}
\item \href{https://otexts.com/fpp2/missing-outliers.html}{Forecasting Principles & Practices: Dealing with missing values and outliers}
}
}
\seealso{
\itemize{
\item Box Cox Transformation: \code{\link[=box_cox_vec]{box_cox_vec()}}
\item Lag Transformation: \code{\link[=lag_vec]{lag_vec()}}
\item Differencing Transformation: \code{\link[=diff_vec]{diff_vec()}}
\item Rolling Window Transformation: \code{\link[=slidify_vec]{slidify_vec()}}
\item Loess Smoothing Transformation: \code{\link[=smooth_vec]{smooth_vec()}}
\item Fourier Series: \code{\link[=fourier_vec]{fourier_vec()}}
\item Missing Value Imputation for Time Series: \code{\link[=ts_impute_vec]{ts_impute_vec()}}
}
}
