% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-broom.R
\name{validate_index}
\alias{validate_index}
\title{Validates data frame has column named the same name as variable index_rename}
\usage{
validate_index(ret, index_rename)
}
\arguments{
\item{ret}{An object of class tibble}

\item{index_rename}{A variable indicating the index name to be used in the
tibble returned}
}
\description{
Validates data frame has column named the same name as variable index_rename
}
