% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wikipedia_traffic_daily.R
\docType{data}
\name{wikipedia_traffic_daily}
\alias{wikipedia_traffic_daily}
\title{Sample Daily Time Series Data from the Web Traffic Forecasting (Wikipedia) Competition}
\format{
A tibble: 9,743 x 3
\itemize{
\item \code{Page} Character. Page information.
\item \code{date} Date. Daily timestamp.
\item \code{value} Numeric. Daily views of the wikipedia article.
}
}
\source{
\itemize{
\item \href{https://www.kaggle.com/c/web-traffic-time-series-forecasting}{Kaggle Competition Website}
}
}
\usage{
wikipedia_traffic_daily
}
\description{
The Kaggle "Web Traffic Forecasting" (Wikipedia) Competition
used \strong{Google Analytics Web Traffic Data} for 145,000 websites.
Each of these time series represent a number of daily views of a different Wikipedia articles.
The competition began July 13th, 2017 and ended November 15th, 2017.
This dataset includes a \strong{Sample of 10 article pages (10 total time series)}.
}
\details{
This is a sample of 10 Daily data sets from the Kaggle Web Traffic Forecasting (Wikipedia) Competition
}
\examples{

wikipedia_traffic_daily

}
\keyword{datasets}
