% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-broom.R
\name{arima_string}
\alias{arima_string}
\title{Print the ARIMA model parameters (refer to forecast:::arima.string)
https://github.com/robjhyndman/forecast/blob/master/R/arima.R}
\usage{
arima_string(object, padding = FALSE)
}
\arguments{
\item{object}{An object of class Arima}

\item{padding}{Add padding to the name returned}
}
\description{
Print the ARIMA model parameters (refer to forecast:::arima.string)
https://github.com/robjhyndman/forecast/blob/master/R/arima.R
}
