% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-time_series.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Simple Interactive Time Series Plotting}
\usage{
plot_time_series(
  .data,
  .date_var,
  .value,
  ...,
  .facet_ncol = 1,
  .facet_scales = "free_y",
  .line_color = "#2c3e50",
  .line_size = 0.5,
  .y_intercept = NULL,
  .y_intercept_color = "#2c3e50",
  .smooth = TRUE,
  .smooth_period = NULL,
  .smooth_degree = 2,
  .smooth_color = "#3366FF",
  .smooth_size = 1,
  .title = "Time Series Plot",
  .x_lab = "",
  .y_lab = "",
  .interactive = TRUE
)
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame}}

\item{.date_var}{A column containing either date or date-time values}

\item{.value}{A column containing numeric values}

\item{...}{One or more grouping columns that broken out into \code{ggplot2} facets.
These can be selected using \code{tidyselect()} helpers (e.g \code{contains()}).}

\item{.facet_ncol}{Facets: Number of facet columns.}

\item{.facet_scales}{Facets: Options include "fixed", "free", "free_y", "free_x"}

\item{.line_color}{Line color. Use keyword: "scale_color" to change the color by the facet.}

\item{.line_size}{Line size}

\item{.y_intercept}{Value for a y-intercept on the plot}

\item{.y_intercept_color}{Color for the y-intercept}

\item{.smooth}{Logical - Whether or not to include a trendline smoother.
Uses See \code{\link[=smooth_vec]{smooth_vec()}} to apply a LOESS smoother.}

\item{.smooth_period}{Number of observations to include in the Loess Smoother.
See \code{\link[=smooth_vec]{smooth_vec()}}.}

\item{.smooth_degree}{Flexibility of Loess Polynomial.
Either 0, 1, 2 (0 = lest flexible, 2 = more flexible).}

\item{.smooth_color}{Smoother line color}

\item{.smooth_size}{Smoother line size}

\item{.title}{Title for the plot}

\item{.x_lab}{X-axis label for the plot}

\item{.y_lab}{Y-axis label for the plot}

\item{.interactive}{Returns either a static (\code{ggplot2}) visualization or an interactive (\code{plotly}) visualization}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
\code{plot_time_series()} returns a time series visualization in either static (\code{ggplot2}) or
interactive (\code{plotly}) formats.
}
\details{
\code{plot_time_series()} is a scalable function that works with both \emph{ungrouped} and \emph{grouped}
\code{data.frame} objects (and \code{tibbles}!).

\strong{Interactive by Default}

\code{plot_time_series()} is built for exploration using:
\itemize{
\item \strong{Interactive Plots:} \code{plotly} (default) - Great for exploring!
\item \strong{Static Plots:} \code{ggplot2} (set \code{.interactive = FALSE}) - Great for PDF Reports
}

By default, an interactive \code{plotly} visualization is returned.

\strong{Scalable with Facets & Dplyr Groups}

\code{plot_time_series()} returns multiple time series plots using \code{ggplot2} facets:
\itemize{
\item \code{group_b()} - If groups are detected, multiple facets are returned
\item \code{plot_time_series(.facets)} - You can manually supply facets as well.
}

Multiple groups are collapsed into one group to return a single facet that is
the distinct combination of multiple groups.
}
\examples{
library(tidyverse)
library(tidyquant)
library(lubridate)
library(timetk)

# Works with individual time series
FANG \%>\%
    filter(symbol == "FB") \%>\%
    plot_time_series(date, adjusted,
                     .facet_ncol = 2, .interactive = FALSE)

# Works with groups
FANG \%>\%
    group_by(symbol) \%>\%
    plot_time_series(date, adjusted,
                     .facet_ncol = 2, .interactive = FALSE)

# Can also group inside
FANG \%>\%
    mutate(year = year(date)) \%>\%
    plot_time_series(date, adjusted,
                     symbol, year, # add groups
                     .facet_ncol   = 4,
                     .facet_scales = "free",
                     .interactive  = FALSE)

# Plotly - Interactive Visualization By Default (Great for Exploration)
FANG \%>\%
    plot_time_series(date, adjusted, symbol)

# ggplot2 - static visualization (Great for PDF Reports)
FANG \%>\%
    plot_time_series(
         date, adjusted, symbol,
        .facet_ncol    = 2,
        .line_color    = "scale_color",
        .smooth_period = 180,
        .interactive   = FALSE) +
   theme_tq_dark() +
   scale_color_viridis_d()


}
