% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-time_series.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Interactive Plotting for One or More Time Series}
\usage{
plot_time_series(
  .data,
  .date_var,
  .value,
  .color_var = NULL,
  .facet_vars = NULL,
  .facet_ncol = 1,
  .facet_scales = "free_y",
  .facet_collapse = TRUE,
  .facet_collapse_sep = " ",
  .line_color = "#2c3e50",
  .line_size = 0.5,
  .line_type = 1,
  .line_alpha = 1,
  .y_intercept = NULL,
  .y_intercept_color = "#2c3e50",
  .smooth = TRUE,
  .smooth_period = "auto",
  .smooth_message = FALSE,
  .smooth_span = NULL,
  .smooth_degree = 2,
  .smooth_color = "#3366FF",
  .smooth_size = 1,
  .smooth_alpha = 1,
  .legend_show = TRUE,
  .title = "Time Series Plot",
  .x_lab = "",
  .y_lab = "",
  .color_lab = "Legend",
  .interactive = TRUE,
  .plotly_slider = FALSE
)
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame} with a time-based column}

\item{.date_var}{A column containing either date or date-time values}

\item{.value}{A column containing numeric values}

\item{.color_var}{A categorical column that can be used to change the
line color}

\item{.facet_vars}{One or more grouping columns that broken out into \code{ggplot2} facets.
These can be selected using \code{tidyselect()} helpers (e.g \code{contains()}).}

\item{.facet_ncol}{Number of facet columns.}

\item{.facet_scales}{Control facet x & y-axis ranges.
Options include "fixed", "free", "free_y", "free_x"}

\item{.facet_collapse}{Multiple facets included on one facet strip instead of
multiple facet strips.}

\item{.facet_collapse_sep}{The separator used for collapsing facets.}

\item{.line_color}{Line color. Overrided if \code{.color_var} is specified.}

\item{.line_size}{Line size.}

\item{.line_type}{Line type.}

\item{.line_alpha}{Line alpha (opacity). Range: (0, 1).}

\item{.y_intercept}{Value for a y-intercept on the plot}

\item{.y_intercept_color}{Color for the y-intercept}

\item{.smooth}{Logical - Whether or not to include a trendline smoother.
Uses See \code{\link[=smooth_vec]{smooth_vec()}} to apply a LOESS smoother.}

\item{.smooth_period}{Number of observations to include in the Loess Smoother.
Set to "auto" by default, which uses \code{tk_get_trend()}
to determine a logical trend cycle.}

\item{.smooth_message}{Logical.
Whether or not to return the trend selected as a message.
Useful for those that want to see what \code{.smooth_period} was selected.}

\item{.smooth_span}{Percentage of observations to include in the Loess Smoother.
You can use either period or span. See \code{\link[=smooth_vec]{smooth_vec()}}.}

\item{.smooth_degree}{Flexibility of Loess Polynomial.
Either 0, 1, 2 (0 = lest flexible, 2 = more flexible).}

\item{.smooth_color}{Smoother line color}

\item{.smooth_size}{Smoother line size}

\item{.smooth_alpha}{Smoother alpha (opacity). Range: (0, 1).}

\item{.legend_show}{Toggles on/off the Legend}

\item{.title}{Title for the plot}

\item{.x_lab}{X-axis label for the plot}

\item{.y_lab}{Y-axis label for the plot}

\item{.color_lab}{Legend label if a \code{color_var} is used.}

\item{.interactive}{Returns either a static (\code{ggplot2}) visualization or an interactive (\code{plotly}) visualization}

\item{.plotly_slider}{If TRUE, returns a plotly date range slider.}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
A workhorse time-series plotting function that generates interactive \code{plotly} plots,
consolidates 20+ lines of \code{ggplot2} code, and scales well to many time series.
}
\details{
\code{plot_time_series()} is a scalable function that works with both \emph{ungrouped} and \emph{grouped}
\code{data.frame} objects (and \code{tibbles}!).

\strong{Interactive by Default}

\code{plot_time_series()} is built for exploration using:
\itemize{
\item \strong{Interactive Plots:} \code{plotly} (default) - Great for exploring!
\item \strong{Static Plots:} \code{ggplot2} (set \code{.interactive = FALSE}) - Great for PDF Reports
}

By default, an interactive \code{plotly} visualization is returned.

\strong{Scalable with Facets & Dplyr Groups}

\code{plot_time_series()} returns multiple time series plots using \code{ggplot2} facets:
\itemize{
\item \code{group_by()} - If groups are detected, multiple facets are returned
\item \code{plot_time_series(.facet_vars)} - You can manually supply facets as well.
}

\strong{Can Transform Values just like ggplot}

The \code{.values} argument accepts transformations just like \code{ggplot2}.
For example, if you want to take the log of sales you can use
a call like \code{plot_time_series(date, log(sales))} and the log transformation
will be applied.

\strong{Smoother Period / Span Calculation}

The \code{.smooth = TRUE} option returns a smoother that is calculated based on either:
\enumerate{
\item A \code{.smooth_period}: Number of observations
\item A \code{.smooth_span}: A percentage of observations
}

By default, the \code{.smooth_period} is automatically calculated using 75\% of the observertions.
This is the same as \code{geom_smooth(method = "loess", span = 0.75)}.

A user can specify a time-based window (e.g. \code{.smooth_period = "1 year"})
or a numeric value (e.g. \code{smooth_period = 365}).

Time-based windows return the median number of observations in a window using \code{tk_get_trend()}.
}
\examples{

library(tidyverse)
library(tidyquant)
library(lubridate)
library(timetk)

# Works with individual time series
FANG \%>\%
    filter(symbol == "FB") \%>\%
    plot_time_series(date, adjusted, .interactive = FALSE)

# Works with groups
FANG \%>\%
    group_by(symbol) \%>\%
    plot_time_series(date, adjusted,
                     .facet_ncol  = 2,     # 2-column layout
                     .interactive = FALSE)

# Can also group inside & use .color_var
FANG \%>\%
    mutate(year = year(date)) \%>\%
    plot_time_series(date, adjusted,
                     .facet_vars   = c(symbol, year), # add groups/facets
                     .color_var    = year,            # color by year
                     .facet_ncol   = 4,
                     .facet_scales = "free",
                     .interactive  = FALSE)

# Can apply transformations to .value or .color_var
# - .value = log(adjusted)
# - .color_var = year(date)
FANG \%>\%
    plot_time_series(date, log(adjusted),
                     .color_var    = year(date),
                     .facet_vars   = contains("symbol"),
                     .facet_ncol   = 2,
                     .facet_scales = "free",
                     .y_lab        = "Log Scale",
                     .interactive  = FALSE)



}
