% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_index.R
\name{sw_index}
\alias{sw_index}
\title{Extract an index of date or datetime from time series}
\usage{
sw_index(data)
}
\arguments{
\item{data}{A time-based tibble or time-series object.}
}
\value{
Returns a vector of date or date times
}
\description{
Extract an index of date or datetime from time series
}
\details{
\code{sw_index} is used to extract the date or datetime index from various
time series data types.
}
\examples{

data_tbl <- tibble::tibble(
    date = seq.Date(from = as.Date("2000-01-01"), by = 1, length.out = 5),
    x    = rnorm(5) * 10,
    y    = 5:1
)
sw_index(data_tbl)

data_ts <- sw_ts(data_tbl)
sw_index(data_ts)

}
\seealso{
\code{\link[=sw_tbl]{sw_tbl()}}, \code{\link[=sw_xts]{sw_xts()}}, \code{\link[=sw_zoo]{sw_zoo()}}, \code{\link[=sw_zooreg]{sw_zooreg()}}, \code{\link[=sw_ts]{sw_ts()}}
}
