% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep-package.R
\docType{package}
\name{sweep_package}
\alias{sweep_package}
\alias{sweep_package-package}
\title{sweep: Time series coercion tools}
\description{
The \code{sweep} package combines a collection of coercion tools for
time series analysis in the "tidyverse".
}
\details{
The \code{sweep} package has two main benefits:
\enumerate{
\item It simplifies the coercion process between time-based tibbles (\code{tbl})
and the major time series data types \code{xts}, \code{zoo}, \code{zooreg}, and \code{ts}
\item It maximizes time-based data retention during coercion to regularized time series
}

To learn more about \code{sweep}, start with the vignettes:
\code{browseVignettes(package = "sweep")}
}
