% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep-package.R
\docType{package}
\name{sweep_package}
\alias{sweep_package}
\alias{sweep_package-package}
\title{sweep: Tidy Tools for Forecasting and Working with Time Series}
\description{
The \code{sweep} package is a swiss army knife for "tidy" time series forecasting.
}
\details{
The main benefit of \code{sweep} is to
enable coercion of time series and \code{forecast} objects between the various data structures
that exist within R such as \code{ts}, \code{mts}, \code{irts}, \code{zoo}, \code{zooreg}, \code{xts}, and
\code{tibble}. Further, time-series-like objects (such as
\code{forecast}, \code{stl}, \code{HoltzWinters}, etc) can be coerced to "tidy" tibbles.

The goal of the \code{sweep} package is to "tidy" up the
forecasting workflow. The \code{sweep} package provides tools to solve three
forecasting / time-series scenarios:
\enumerate{
\item It provides functions to easily coerce between
tibbles ("tidy" data frames) and the various time series data structures.
Refer to \code{\link[=sw_tbl]{sw_tbl()}}, \code{\link[=sw_xts]{sw_xts()}}, \code{\link[=sw_zoo]{sw_zoo()}}, and \code{\link[=sw_ts]{sw_ts()}}.
\item It extends the \code{broom} package by providing \code{\link[=tidy]{tidy()}}, \code{\link[=glance]{glance()}}, and
\code{\link[=augment]{augment()}} functions for various models (such as
\code{stl}, \code{HoltzWinters}, \code{tbats}, etc).
\item It provides the \code{\link[=sweep]{sweep()}} function which coerces
\code{forecast} objects to tibbles for easy visualization.
}

To learn more about \code{sweep}, start with the vignettes:
\code{browseVignettes(package = "sweep")}
}
