% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetk-package.R
\docType{package}
\name{timetk}
\alias{timetk}
\alias{timetk-package}
\title{timetk: Time Series Analysis in the Tidyverse}
\description{
The \code{timetk} package combines a collection of coercion tools for
time series analysis.
}
\details{
The \code{timetk} package has several benefits:
\enumerate{
\item Visualizing Time Series
\item Wrangling Time Series.
\item Preprocessing and Feature Engineering.
}

To learn more about \code{timetk}, start with the documentation:
\url{https://business-science.github.io/timetk/}
}
