% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-tk_anomaly_diagnostics.R
\name{tk_anomaly_diagnostics}
\alias{tk_anomaly_diagnostics}
\alias{tk_anomaly_diagnostics.data.frame}
\title{Group-wise STL Decomposition (Season, Trend, Remainder)}
\usage{
tk_anomaly_diagnostics(
  .data,
  .date_var,
  .value,
  .frequency = "auto",
  .trend = "auto",
  .alpha = 0.05,
  .max_anomalies = 0.2,
  .message = TRUE
)

\method{tk_anomaly_diagnostics}{data.frame}(
  .data,
  .date_var,
  .value,
  .frequency = "auto",
  .trend = "auto",
  .alpha = 0.05,
  .max_anomalies = 0.2,
  .message = TRUE
)
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame} with a time-based column}

\item{.date_var}{A column containing either date or date-time values}

\item{.value}{A column containing numeric values}

\item{.frequency}{Controls the seasonal adjustment (removal of seasonality).
Input can be either "auto", a time-based definition (e.g. "2 weeks"),
or a numeric number of observations per frequency (e.g. 10).
Refer to \code{\link[=tk_get_frequency]{tk_get_frequency()}}.}

\item{.trend}{Controls the trend component.
For STL, trend controls the sensitivity of the LOESS smoother, which is used to remove the remainder.
Refer to \code{\link[=tk_get_trend]{tk_get_trend()}}.}

\item{.alpha}{Controls the width of the "normal" range. Lower values are more conservative
while higher values are less prone to incorrectly classifying "normal" observations.}

\item{.max_anomalies}{The maximum percent of anomalies permitted to be identified.}

\item{.message}{A boolean. If \code{TRUE}, will output information related to automatic frequency
and trend selection (if applicable).}
}
\value{
A \code{tibble} or \code{data.frame} with Observed, Season, Trend, Remainder,
and Seasonally-Adjusted features
}
\description{
\code{tk_anomaly_diagnostics()} is the preprocessor for \code{plot_stl_diagnostics()}.
It helps by automating frequency and trend selection.
}
\details{
TODO
}
\examples{
library(dplyr)
library(timetk)


# TODO

}
