% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-tk_augment_roll_apply.R
\name{tk_augment_roll_apply}
\alias{tk_augment_roll_apply}
\title{Augment columns with rolling windows to the data}
\usage{
tk_augment_roll_apply(
  .data,
  .column,
  .period,
  .f,
  ...,
  .align = c("center", "left", "right"),
  .partial = FALSE,
  .names = paste0("roll_apply_", .period)
)
}
\arguments{
\item{.data}{A tibble.}

\item{.column}{A column to have a rolling window transformation applied}

\item{.period}{One or more periods for the rolling window(s)}

\item{.f}{A summary \verb{[function / formula]},}

\item{...}{Optional arguments for the summary function}

\item{.align}{Rolling functions generate \code{.period - 1} fewer values than the incoming vector.
Thus, the vector needs to be aligned. Select one of "center", "left", or "right".}

\item{.partial}{.partial Should the moving window be allowed to return partial (incomplete) windows instead of \code{NA} values.
Set to FALSE by default, but can be switched to TRUE to remove \code{NA}'s.}

\item{.names}{A vector of names for the new columns. Must be of same length as \code{.period}.}
}
\value{
Returns a \code{tibble} object describing the timeseries.
}
\description{
Augment columns with rolling windows to the data
}
\details{
\code{tk_augment_roll_apply()} scales the \code{\link[=roll_apply_vec]{roll_apply_vec()}} function to multiple
time series \code{.periods}. See \code{\link[=roll_apply_vec]{roll_apply_vec()}} for examples and usage of the core function
arguments.
}
\examples{
library(tidyverse)
library(tidyquant)
library(timetk)

FANG \%>\%
    select(symbol, date, adjusted) \%>\%
    group_by(symbol) \%>\%
    tk_augment_roll_apply(
        .column  = adjusted,
        # Multiple rolling windows
        .period  = c(10, 30),
        .f       = AVERAGE,
        .partial = TRUE,
        .names   = str_c("MA_", c(10, 30))
    )

}
\seealso{
\itemize{
\item \code{\link[=roll_apply_vec]{roll_apply_vec()}} - The underlying function that powers \code{tk_augment_roll_apply()}
\item More Augment Operations: \code{\link{tk_augment_timeseries_signature}}
}
}
